+++ exporters/otlp/src/otlp_grpc_metric_exporter_options.cc
@@ -48,10 +48,24 @@ OtlpGrpcMetricExporterOptions::OtlpGrpcMetricExporterOptions()
 }
 
 OtlpGrpcMetricExporterOptions::OtlpGrpcMetricExporterOptions(void *)
-    : aggregation_temporality(PreferredAggregationTemporality::kCumulative)
+    : OtlpGrpcClientOptions(nullptr),
+      aggregation_temporality(PreferredAggregationTemporality::kCumulative)
 {
   use_ssl_credentials = true;
   max_threads         = 0;
+
+#ifdef ENABLE_ASYNC_EXPORT
+  max_concurrent_requests = 64;
+#endif
+}
+
+OtlpGrpcMetricExporterOptions::OtlpGrpcMetricExporterOptions(
+    const OtlpGrpcClientOptions &client_options)
+    : OtlpGrpcClientOptions(client_options),
+      aggregation_temporality(PreferredAggregationTemporality::kCumulative)
+{
+  timeout  = GetOtlpDefaultMetricsTimeout();
+  metadata = GetOtlpDefaultMetricsHeaders();
 
 #ifdef ENABLE_ASYNC_EXPORT
   max_concurrent_requests = 64;